requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
python-multipart
pypdf>=4.0
openai
google-auth
google-auth-oauthlib
google-api-python-client
psycopg2-binary
pgvector